Rename EnrolmentAuthenticationNotification as SignupAuthenticationNotification
