+++ src/domain/attendanceList/AttendanceListPage.tsx
@@ -14,12 +14,12 @@ import Container from '../app/layout/container/Container';
 import MainContent from '../app/layout/mainContent/MainContent';
 import PageWrapper from '../app/layout/pageWrapper/PageWrapper';
 import { SIGNUP_ACTIONS } from '../enrolment/constants';
-import useRegistrationAndEventData from '../enrol../enrolment/enrolmentAuthenticationNotification/EnrolmentAuthenticationNotification
+import useRegistrationAndEventData from '../enrolment/hooks/useRegistrationAndEventData';
 import NotFound from '../notFound/NotFound';
 import { REGISTRATION_INCLUDES } from '../registration/constants';
 import RegistrationInfo from '../registration/registrationInfo/RegistrationInfo';
 import { getRegistrationFields } from '../registration/utils';
-import EnrolmentAuthenticationNotification from '../signup/enrolmentAuthenticationNotification/EnrolmentAuthenticationNotification';
+import SignupAuthenticationNotification from '../signup/signupAuthenticationNotification/SignupAuthenticationNotification';
 import styles from './attendanceListPage.module.scss';
 import AttendeeList from './attendeeList/AttendeeList';
 
@@ -67,7 +67,7 @@ const AttendanceListPage: React.FC<AttendanceListPageProps> = ({
               { active: true, label: t(`attendanceListPage.title`) },
             ]}
           />
-          <EnrolmentAuthenticationNotification
+          <SignupAuthenticationNotification
             action={SIGNUP_ACTIONS.VIEW}
             registration={registration}
           />
